pyproject.toml
[tool.poetry]
name = "blog-generator"
version = "0.1.0"
description = ""
authors = ["Aleksey Petrunnik"]

[tool.poetry.dependencies]
python = "^3.7"
markdown = "^3.2"
pygments = "^2.4"
pymdown-extensions = "^7.1"
jinja2 = "^2.10"
gitpython = "^3.1"

[tool.poetry.dev-dependencies]
pytest = "=5.4.2"
pytest-datadir = "=1.3.1"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
